+++ src/main/java/fei/upce/st60982/pocketbook/Repositories/TaskDAO.java
@@ -3,10 +3,11 @@ package fei.upce.st60982.pocketbook.Repositories;
 import fei.upce.st60982.pocketbook.DataClasses.Task;
 import fei.upce.st60982.pocketbook.DataClasses.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface TaskDAO extends JpaRepository<Task, Long> {
-    Task findTaskById(long id);
+public interface TaskDAO extends JpaRepository<Task, Integer> {
+    Task findTaskById(int id);
     List<Task> findTasksByAuthor(User user);
 }
